refactor(plan_SERVERSPLIT slice 6): the network-debugging bodies leave server.py
The ten `network-debugging` tool bodies move VERBATIM out of
`embedded/server.py` into `embedded/tool_sections/network_debugging.py`. No
behaviour change, no rename, no tool added or removed: the served surface is
byte-identical to `tests/goldens/tool_surface.json`, and all three loaded
identities of `server.py` still build a full 94-tool app.

Two things make this slice different from the five before it.

`_CAPTURE_OFF_NOTE` moves WITH the section. It is the first module constant the
migration relocates: in `server.py` it sat wedged between `get_request_details`
and `get_response_details` — the plan's baseline table names it as a constant
embedded mid-section — and it now sits at the top of the module whose three
`capture_note` tools are its only readers. The string is byte-identical; only its
position moved.

And this is the plan's heaviest raw-`.fn` test coupling:
`tests/test_server_network_tools.py` reaches fifteen call sites as
`server.<tool>.fn(...)` plus two `server.search_network_requests.description`
reads. The census was re-run for this slice rather than assumed from slices 1-5,
and every one of those is a module-attribute read `server.py`'s binding loop
still satisfies, so none needed re-pointing. No import was pruned either:
`server.py` still drives `network_interceptor` from `spawn_browser`,
`close_instance` and two `@mcp.resource` handlers.

server.py: 2660 -> 2391 LOC; the GRANDFATHER cap ratchets DOWN to that measured
post-`ruff format` actual, and `tests/source_scan.py`'s floor ratchets UP 7 -> 8.
`SECTION_MODULES` gains a section that is canonically FIRST, so the mid-migration
`SECTION_TOOLS` key order shifts again — verified harmless a second time
(`gen_release_contract --check` clean, `--list-sections` totals only,
`release_evidence` sorts, the wire golden is keyed by tool name).

Verified: unit lane 2332 passed / 1 skipped (unchanged from slice 5 — nothing
pruned this time); wire surface IDENTICAL; three-identity load SECTION_TOOLS 94 /
per-app [94, 94, 94]; `--list-sections` subprocess "Total: 94 tools"; ruff
format+check, budgets, suppression owners, vulture and
`gen_release_contract --check` clean; ty exit 0 at 80 warnings, two more than
slice 5 — the two carried `response.dict()` / `request.dict()` pydantic-v1 calls,
visible only because `server.py` is in ty's exclude list while a section module
is not, and a behaviour change to narrow. Real Chrome, one file at a time:
`test_e2e_network_capture_shape.py` 1 passed, `test_e2e_data_tools.py` 5 passed,
`test_e2e_extra_headers.py` 2 passed; real stdio `test_e2e_transport.py` 1 passed.

The plan's execution log gains the slices 4-6 section with eight drift rows.
