+++ src/stealth_chrome_devtools_mcp/embedded/server.py
@@ -1086,275 +1086,6 @@ async def take_screenshot(
             os.unlink(tmp_path)
 
 
-@section_tool("network-debugging")
-async def list_network_requests(
-    instance_id: str, filter_type: str | None = None
-) -> list[dict[str, Any]] | dict[str, Any]:
-    """
-    List captured network requests.
-
-    Browser-internal traffic (chrome://, chrome-extension://, devtools://, about:)
-    is excluded from capture by default — see set_network_capture_filters.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        filter_type (Optional[str]): Filter by CDP resource type, matched
-            case-insensitively (e.g., 'document', 'image', 'script', 'xhr').
-
-    Returns:
-        Union[List[Dict[str, Any]], Dict[str, Any]]: List of network requests, or file metadata if response too large.
-    """
-    requests = await network_interceptor.list_requests(instance_id, filter_type)
-    formatted_requests = [
-        {
-            "request_id": req.request_id,
-            "url": req.url,
-            "method": req.method,
-            "resource_type": req.resource_type,
-            "timestamp": req.timestamp.isoformat(),
-        }
-        for req in requests
-    ]
-
-    return response_handler.handle_response(formatted_requests, "network_requests")
-
-
-@section_tool("network-debugging")
-async def get_request_details(request_id: str) -> dict[str, Any] | None:
-    """
-    Get detailed information about a network request.
-
-    Args:
-        request_id (str): Network request ID.
-
-    Returns:
-        Optional[Dict[str, Any]]: Request details including headers, cookies, and body.
-    """
-    request = await network_interceptor.get_request(request_id)
-    if request:
-        return request.dict()
-    return None
-
-
-# Surfaced by the body-consuming tools when a body is absent because capture is
-# off, so an empty body doesn't read as a broken tool (F-605, off-by-default).
-_CAPTURE_OFF_NOTE = (
-    "response-body capture is off; enable via "
-    "set_network_capture_filters(capture_bodies=True) or "
-    "STEALTH_MCP_NETWORK_CAPTURE_BODIES=1"
-)
-
-
-@section_tool("network-debugging")
-async def get_response_details(request_id: str) -> dict[str, Any] | None:
-    """
-    Get response details for a network request.
-
-    Response bodies are only stored when capture is enabled (off by default);
-    status/headers/content-type metadata is always captured. When the body is
-    absent because capture is off, a ``capture_note`` explains how to enable it.
-
-    Args:
-        request_id (str): Network request ID.
-
-    Returns:
-        Optional[Dict[str, Any]]: Response details including status, headers, and metadata.
-    """
-    response = await network_interceptor.get_response(request_id)
-    if response:
-        result = response.dict()
-        if response.body is None and not get_settings().network_capture_bodies:
-            result["capture_note"] = _CAPTURE_OFF_NOTE
-        return result
-    return None
-
-
-@section_tool("network-debugging")
-async def get_response_content(instance_id: str, request_id: str) -> str | None:
-    """
-    Get response body content.
-
-    Live-refetches the body from CDP on demand, so it works even when
-    response-body capture is off (the default) and independent of the stored-body
-    cap. CDP evicts bodies from its own buffer quickly, so fetch soon after load.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        request_id (str): Network request ID.
-
-    Returns:
-        Optional[str]: Response body as text (base64 encoded for binary).
-    """
-    tab = await _require_tab(browser_manager, instance_id)
-    body = await _with_cdp_timeout(
-        network_interceptor.get_response_body(tab, request_id), instance_id=instance_id
-    )
-    if body:
-        try:
-            return body.decode("utf-8")
-        except UnicodeDecodeError:
-            import base64
-
-            return base64.b64encode(body).decode("utf-8")
-    return None
-
-
-@section_tool("network-debugging")
-async def search_network_requests(
-    instance_id: str,
-    url_pattern: str | None = None,
-    method: str | None = None,
-    status_code: int | None = None,
-    response_contains: str | None = None,
-    payload_contains: str | None = None,
-    resource_type: str | None = None,
-    limit: int = 100,
-    offset: int = 0,
-) -> dict[str, Any]:
-    """
-    Search network requests; ``url_pattern`` filters the URL (substring, not glob).
-
-    Args:
-        instance_id (str): Browser instance ID.
-        url_pattern (Optional[str]): URL filter — case-insensitive substring.
-        method (Optional[str]): HTTP method, matched whole, case-insensitively.
-        status_code (Optional[int]): Exact response status code.
-        response_contains (Optional[str]): Case-insensitive substring of the body.
-        payload_contains (Optional[str]): Case-insensitive substring of the payload.
-        resource_type (Optional[str]): Case-insensitive substring of the CDP type.
-        limit (int): Max results per page.
-        offset (int): Starting index for pagination.
-
-    Returns:
-        Dict[str, Any]: Paginated results with metadata. Includes a ``capture_note``
-        when body capture is off (response_contains matching is unavailable then).
-    """
-    result = await network_interceptor.search_requests(
-        instance_id,
-        url_pattern,
-        method,
-        status_code,
-        response_contains,
-        payload_contains,
-        resource_type,
-        limit,
-        offset,
-    )
-    filters = await network_interceptor.get_capture_filters(instance_id)
-    if not filters.get("capture_bodies"):
-        result["capture_note"] = _CAPTURE_OFF_NOTE
-    return result
-
-
-@section_tool("network-debugging")
-async def export_network_data(instance_id: str, filepath: str) -> dict[str, Any]:
-    """
-    Export network data to JSON file.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        filepath (str): Path to save JSON file.
-
-    Returns:
-        Dict[str, Any]: ``{"success": bool}``, plus a ``capture_note`` when body
-        capture is off (exported responses have no bodies until it is enabled).
-    """
-    success = await network_interceptor.export_to_json(instance_id, filepath)
-    result: dict[str, Any] = {"success": success}
-    filters = await network_interceptor.get_capture_filters(instance_id)
-    if not filters.get("capture_bodies"):
-        result["capture_note"] = _CAPTURE_OFF_NOTE
-    return result
-
-
-@section_tool("network-debugging")
-async def import_network_data(instance_id: str, filepath: str) -> bool:
-    """
-    Import network data from JSON file.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        filepath (str): Path to JSON file.
-
-    Returns:
-        bool: True if successful.
-    """
-    return await network_interceptor.import_from_json(instance_id, filepath)
-
-
-@section_tool("network-debugging")
-async def set_network_capture_filters(
-    instance_id: str,
-    include_types: list[str] | None = None,
-    exclude_types: list[str] | None = None,
-    capture_bodies: bool | None = None,
-    capture_internal_urls: bool | None = None,
-) -> bool:
-    """
-    Set resource type filters for network capture to reduce memory usage.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        include_types (Optional[List[str]]): Only capture these types (e.g., ['XHR', 'Fetch', 'Document']).
-        exclude_types (Optional[List[str]]): Exclude these types (e.g., ['Image', 'Stylesheet', 'Font', 'Script']).
-        capture_bodies (Optional[bool]): Enable/disable response-body capture for this
-            instance (default off; overrides STEALTH_MCP_NETWORK_CAPTURE_BODIES). Each
-            argument is merged — passing only capture_bodies keeps include/exclude.
-        capture_internal_urls (Optional[bool]): Capture Chrome's own traffic —
-            chrome://, chrome-extension://, chrome-error://, devtools://, about: —
-            which is EXCLUDED by default so the page's requests are not drowned out
-            (overrides STEALTH_MCP_NETWORK_CAPTURE_INTERNAL_URLS).
-
-    Type matching is case-insensitive, so 'document' and 'Document' behave alike.
-    Common resource types: Document, Stylesheet, Image, Media, Font, Script, XHR, Fetch, WebSocket, Manifest, Other
-
-    Returns:
-        bool: True if successful.
-    """
-    await network_interceptor.set_capture_filters(
-        instance_id,
-        include_types,
-        exclude_types,
-        capture_bodies,
-        capture_internal_urls,
-    )
-    return True
-
-
-@section_tool("network-debugging")
-async def get_network_capture_filters(instance_id: str) -> dict[str, Any]:
-    """
-    Get current network capture filters plus resolved body-capture state.
-
-    Args:
-        instance_id (str): Browser instance ID.
-
-    Returns:
-        Dict[str, Any]: 'include'/'exclude' lists, the resolved 'capture_bodies'
-        and 'capture_internal_urls' flags, and body-store usage
-        ('body_store_bytes', 'body_store_max_bytes', 'body_max_bytes').
-    """
-    return await network_interceptor.get_capture_filters(instance_id)
-
-
-@section_tool("network-debugging")
-async def modify_headers(instance_id: str, headers: dict[str, str]) -> bool:
-    """
-    Modify request headers for future requests.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        headers (Dict[str, str]): Headers to add/modify.
-
-    Returns:
-        bool: True if modified successfully.
-    """
-    tab = await _require_tab(browser_manager, instance_id)
-    return await _with_cdp_timeout(
-        network_interceptor.modify_headers(tab, headers), instance_id=instance_id
-    )
-
-
 @mcp.resource("browser://{instance_id}/state")
 async def get_browser_state_resource(instance_id: str) -> str:
     """
